+++ seeds/users.json
@@ -16,5 +16,11 @@
         "email": "example@mail.com",
         "password": "$2a$10$ldmD9tWkqsxsc4/n2zlf6OiXB3bjzoKun/P6uoqmyW5cte5tyr6ty",
         "isAdmin": false
+    },
+    {
+        "name": "Lorem Ipsum",
+        "email": "1",
+        "password": "$2a$10$ldmD9tWkqsxsc4/n2zlf6OiXB3bjzoKun/P6uoqmyW5cte5tyr6ty",
+        "isAdmin": true
     }
 ]
