+++ src/frontend/presentational/components/login.tsx
@@ -9,7 +9,7 @@ import IconButton from 'material-ui/IconButton';
 import Visibility from 'material-ui-icons/Visibility';
 import VisibilityOff from 'material-ui-icons/VisibilityOff';  
 
-export class Login extends React.Component<any, any> {
+export class LoginForm extends React.Component<any, any> {
   handleChange: (name: any) => (event: any) => void;
   constructor(props) {
     super(props);
